Made things into doubles for type safety
:(

+++ src/org/alixia/libs/evaluator/api/terms/FactorialTermWrapper.java
@@ -2,7 +2,7 @@ package org.alixia.libs.evaluator.api.terms;
 
 import org.alixia.libs.evaluator.api.functions.SimpleFunction;
 
-public class FactorialTermWrapper implements Term<Long> {
+public class FactorialTermWrapper implements Term<Double> {
 
 	private final Term<?> term;
 
@@ -11,9 +11,9 @@ public class FactorialTermWrapper implements Term<Long> {
 	}
 
 	@Override
-	public Long evaluate() {
+	public Double evaluate() {
 		try {
-			return SimpleFunction.FACTORIAL.evaluate(((java.lang.Number) term.evaluate()).longValue()).evaluate();
+			return SimpleFunction.FACTORIAL.evaluate(((java.lang.Number) term.evaluate())).evaluate();
 		} catch (ClassCastException e) {
 			throw new RuntimeException("The factorial operator was illegally applied to a term. Term: " + term
 					+ ", Term Value: " + term.evaluate());
